fix(admin): move account management to users and replace by admin dashboard for statistics
